ex 1: removed new `-1` and added a note
Iterating from 0 to `len(L)-1` prevents an error but not all list items
are added up if `len(L) != 2^n`. Thus, an error was fixed but the code
yields a wrong result. Fixing the code makes it more complex which is
not reasonable because we want to have a simple example. Hence, the best
solution is to remove `-1` at `len(L)-1` again and add a note indicating
the issue.